Refactor to let player choose with buttons
Previously, the user had to type in his choice, and the program played
exactly five rounds. The "Revisiting Rock Paper Scissors" lesson in
The Odin Project instructed me to insert three buttons to let the user
make his choice. Upon the press of any button, the program simulates
a round. The game ends when either the player or computer reaches five
points.